resources: move loader classes to their own directory
