--- src/renderer/cache.hpp
@@ -1,177 +1,0 @@
-#pragma once
-
-#include <bimg/decode.h>
-
-#include <entt/core/hashed_string.hpp>
-#include <entt/resource/cache.hpp>
-#include <memory>
-#include <string>
-#include <unordered_map>
-#include <vector>
-
-#include "bgfx/bgfx.h"
-#include "bgfx/defines.h"
-#include "bimg/bimg.h"
-#include "bx/allocator.h"
-#include "bx/bx.h"
-#include "bx/file.h"
-#include "core/sys/log.hpp"
-#include "core/sys/mem.hpp"
-#include "renderer/asset.hpp"
-#include "renderer/material.hpp"
-#include "renderer/model_loader.hpp"
-#include "renderer/primitive.hpp"
-#include "renderer/shader.hpp"
-
-using namespace entt::literals;
-
-#define WATO_TEXTURE_CACHE (ResourceCache::Instance().TextureCache)
-#define WATO_PROGRAM_CACHE (ResourceCache::Instance().ShaderCache)
-#define WATO_MODEL_CACHE   (ResourceCache::Instance().ModelCache)
-
-inline static void imageReleaseCb(void* aPtr, void* aUserData)
-{
-    BX_UNUSED(aPtr);
-    auto* imageContainer = (bimg::ImageContainer*)aUserData;
-    bimg::imageFree(imageContainer);
-}
-
-struct TextureLoader final {
-    using result_type = std::shared_ptr<bgfx::TextureHandle>;
-
-    result_type operator()(
-        const char*              aName,
-        uint64_t                 aFlags       = BGFX_TEXTURE_NONE | BGFX_SAMPLER_NONE,
-        uint8_t                  aSkip        = 0,
-        bgfx::TextureInfo*       aInfo        = nullptr,
-        bimg::Orientation::Enum* aOrientation = nullptr)
-    {
-        BX_UNUSED(aSkip);
-        bgfx::TextureHandle handle = BGFX_INVALID_HANDLE;
-
-        uint32_t    size      = 0;
-        std::string assetPath = FindAsset(aName);
-        if (assetPath == "") {
-            throw std::runtime_error(fmt::format("cannot find asset {}", aName));
-        }
-
-        void* data = load(&mfr, &mallocator, assetPath.c_str(), &size);
-        if (nullptr != data) {
-            bimg::ImageContainer* imageContainer = bimg::imageParse(&mallocator, data, size);
-
-            if (nullptr != imageContainer) {
-                if (nullptr != aOrientation) {
-                    *aOrientation = imageContainer->m_orientation;
-                }
-
-                const bgfx::Memory* mem = bgfx::makeRef(
-                    imageContainer->m_data,
-                    imageContainer->m_size,
-                    imageReleaseCb,
-                    imageContainer);
-                bx::free(&mallocator, data);
-
-                if (imageContainer->m_cubeMap) {
-                    handle = bgfx::createTextureCube(
-                        uint16_t(imageContainer->m_width),
-                        1 < imageContainer->m_numMips,
-                        imageContainer->m_numLayers,
-                        bgfx::TextureFormat::Enum(imageContainer->m_format),
-                        aFlags,
-                        mem);
-                } else if (1 < imageContainer->m_depth) {
-                    handle = bgfx::createTexture3D(
-                        uint16_t(imageContainer->m_width),
-                        uint16_t(imageContainer->m_height),
-                        uint16_t(imageContainer->m_depth),
-                        1 < imageContainer->m_numMips,
-                        bgfx::TextureFormat::Enum(imageContainer->m_format),
-                        aFlags,
-                        mem);
-                } else if (bgfx::isTextureValid(
-                               0,
-                               false,
-                               imageContainer->m_numLayers,
-                               bgfx::TextureFormat::Enum(imageContainer->m_format),
-                               aFlags)) {
-                    handle = bgfx::createTexture2D(
-                        uint16_t(imageContainer->m_width),
-                        uint16_t(imageContainer->m_height),
-                        1 < imageContainer->m_numMips,
-                        imageContainer->m_numLayers,
-                        bgfx::TextureFormat::Enum(imageContainer->m_format),
-                        aFlags,
-                        mem);
-                }
-
-                if (bgfx::isValid(handle)) {
-                    DBG("Loaded texture {}", assetPath);
-                    bgfx::setName(handle, assetPath.c_str());
-                }
-
-                if (nullptr != aInfo) {
-                    bgfx::calcTextureSize(
-                        *aInfo,
-                        uint16_t(imageContainer->m_width),
-                        uint16_t(imageContainer->m_height),
-                        uint16_t(imageContainer->m_depth),
-                        imageContainer->m_cubeMap,
-                        1 < imageContainer->m_numMips,
-                        imageContainer->m_numLayers,
-                        bgfx::TextureFormat::Enum(imageContainer->m_format));
-                }
-            }
-        }
-
-        return std::make_shared<bgfx::TextureHandle>(handle);
-    }
-
-   private:
-    bx::DefaultAllocator mallocator;
-    bx::FileReader       mfr;
-};
-
-struct ProgramLoader final {
-    using uniform_desc_map = std::unordered_map<std::string, UniformDesc>;
-    using result_type      = std::shared_ptr<Shader>;
-
-    result_type
-    operator()(const char* aVsName, const char* aFsName, const uniform_desc_map& aUniforms)
-    {
-        auto handle = loadProgram(&mfr, aVsName, aFsName);
-
-        auto uniformHandles = Shader::uniform_map{};
-        for (auto&& [name, desc] : aUniforms) {
-            DBG("creating uniform '{}'", name);
-            uniformHandles[name] = bgfx::createUniform(name.c_str(), desc.Type, desc.Number);
-        }
-
-        return std::make_shared<Shader>(handle, uniformHandles);
-    }
-
-   private:
-    bx::FileReader mfr;
-};
-
-using TexCache   = entt::resource_cache<bgfx::TextureHandle, TextureLoader>;
-using ShCache    = entt::resource_cache<Shader, ProgramLoader>;
-using ModelCache = entt::resource_cache<Model, ModelLoader>;
-
-struct ResourceCache {
-    static ResourceCache& Instance()
-    {
-        static ResourceCache rc;
-
-        return rc;
-    }
-
-    ResourceCache(ResourceCache const&)  = delete;
-    void operator=(ResourceCache const&) = delete;
-
-    TexCache     TextureCache;
-    ShCache      ShaderCache;
-    ::ModelCache ModelCache;
-
-   private:
-    ResourceCache() {};
-};
