+++ src/resource/asset.cpp
