+++ src/resource/asset.hpp
